pyproject.toml
[project]
name = "sprintsync"
version = "0.1.0"
description = ""
authors = [
    {name = "aneesarom",email = "speaktoanees@gmail.com"}
]
requires-python = ">=3.11,<3.12"
dependencies = [
    "fastapi (>=0.129.0,<0.130.0)",
    "uvicorn (>=0.41.0,<0.42.0)",
    "supabase (>=2.28.0,<3.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "bcrypt (==3.2.2)",
    "pydantic[email] (>=2.12.5,<3.0.0)",
    "langchain-google-genai (>=4.2.1,<5.0.0)",
    "langchain (>=1.2.10,<2.0.0)",
    "langchain-groq (>=1.1.2,<2.0.0)",
    "structlog (>=25.5.0,<26.0.0)",
    "prometheus-client (>=0.24.1,<0.25.0)",
    "pytest (>=9.0.2,<10.0.0)",
]

[tool.poetry]
package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
